Update to new version some dependencies & connect firebase Swap names of sign_in & sign_up folders and files in
