+++ app/src/main/java/com/example/chessgo/frontend/registration/sign_in/SignInActivity.kt
@@ -1,7 +1,6 @@
 package com.example.chessgo.frontend.registration.sign_in
 
-
-import android.content.ContentValues.TAG
+import android.content.ContentValues
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
@@ -23,111 +22,86 @@ import androidx.compose.ui.text.*
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.chessgo.frontend.MainActivity
-import com.example.chessgo.frontend.registration.sign_in.SignInActivity
 import com.example.chessgo.frontend.registration.sign_up.SignUpActivity
 import com.example.chessgo.ui.theme.ChessgoTheme
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.auth.UserProfileChangeRequest
 import com.google.firebase.auth.ktx.auth
-import com.google.firebase.database.DatabaseReference
-import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
 
-class SignInActivity : ComponentActivity() {
+class SignInActivity:  ComponentActivity() {
+
 
     private lateinit var auth: FirebaseAuth
-    private lateinit var database: DatabaseReference
     private var uiState by mutableStateOf(SignInUiState())
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         auth = Firebase.auth
-        database = Firebase.database.reference
+
     }
 
     override fun onStart() {
         super.onStart()
         setContent {
             ChessgoTheme {
-                RegistrationForm {
-                    val intent = Intent(applicationContext, SignUpActivity::class.java)
+                LoginForm {
+                    val intent = Intent(applicationContext, SignUpActivity::class.java).apply {
+                        putExtra("registration", false)
+                    }
+
                     startActivity(intent)
                 }
             }
         }
     }
 
-    private fun createAccount() {
+    private fun login() {
         if (uiState.isNotEmpty()) {
 
-            auth.createUserWithEmailAndPassword(uiState.email, uiState.password)
+            auth.signInWithEmailAndPassword(uiState.email, uiState.password)
                 .addOnCompleteListener(this) { task ->
                     if (task.isSuccessful) {
                         // Sign in success, update UI with the signed-in user's information
-                        Log.d(TAG, "createUserWithEmail:success")
+                        Log.d(ContentValues.TAG, "createUserWithEmail:success")
                         val user = auth.currentUser
-                        val profileUpdates = UserProfileChangeRequest.Builder()
-                            .setDisplayName(uiState.userName)
-                            .build()
 
-                        updateUI(user, profileUpdates)
+                        updateUI(user)
                     } else {
                         // If sign in fails, display a message to the user.
-                        Log.w(TAG, "createUserWithEmail:failure", task.exception)
+                        Log.w(ContentValues.TAG, "createUserWithEmail:failure", task.exception)
                         Toast.makeText(
                             baseContext,
                             "Authentication failed.",
                             Toast.LENGTH_SHORT,
                         ).show()
-                        updateUI(null, null)
+                        updateUI(null)
                     }
                 }
         }
     }
-    private fun saveUserToDatabase(userName: String?, email: String?, uid: String?){
 
-        val user = User(userName, email)
-        if (uid != null) {
-            System.out.println(1)
-            database.child("Users").child(uid).setValue(user)
-                .addOnSuccessListener {
-                    System.out.println(2)
-                    Log.d(TAG, "User data saved to database")
-                }
-                .addOnFailureListener { e ->
-                    System.out.println(3)
-                    Log.e(TAG, "Error saving user data to database: $e")
-                }
-        }
-    }
-    private fun updateUI(user: FirebaseUser?, profileUpdates: UserProfileChangeRequest?) {
-        profileUpdates?.let {
-            user?.updateProfile(it)?.addOnCompleteListener { profileUpdateTask ->
-                if (profileUpdateTask.isSuccessful) {
-                    System.out.println(uiState.userName)
-                    System.out.println( uiState.email)
-                    System.out.println( auth.currentUser?.uid)
-                    saveUserToDatabase(uiState.userName, uiState.email, auth.currentUser?.uid)
-                    val intent = Intent(applicationContext, MainActivity::class.java)
-                    startActivity(intent)
-                    finish()
-                }
-            }
+    private fun updateUI(user: FirebaseUser?) {
+        if(user != null){
+            val intent = Intent(applicationContext, MainActivity::class.java)
+            startActivity(intent)
+            finish()
         }
     }
 
-    private fun checkIfStringLegal(email: String?, password: String?, userName: String?): Boolean {
-
-        if (email != null && password != null && userName != null) {
-            return false
-        }
-        return true
+    private fun getPhotoUrl(): String? {
+        val user = auth.currentUser
+        return user?.photoUrl?.toString()
     }
 
+    private fun getUserName(): String? {
+        val user = auth.currentUser
+        return if (user != null) {
+            user.displayName
+        } else "Failed to display an user's name"
+    }
 
     @Composable
-    fun RegistrationForm(onSignInClick: () -> Unit){
+    fun LoginForm(onSignUpClick: () -> Unit) {
 
         Column(
             modifier = Modifier.fillMaxSize(),
@@ -135,31 +109,21 @@ class SignInActivity : ComponentActivity() {
             verticalArrangement = Arrangement.Center
         ) {
 
-            Text(text = "Registration", style = TextStyle(fontSize = 40.sp), modifier = Modifier.padding(0.dp, 0.dp, 0.dp, 0.dp))
-
-            Spacer(modifier = Modifier.height(20.dp))
             //var email by remember { mutableStateOf("") }
             TextField(
                 value = uiState.email,
-                onValueChange = {email ->  uiState = uiState.copy(email = email) },
-                label = { Text(text = "Email address") }
-            )
-
-            Spacer(modifier = Modifier.height(20.dp))
-            //var userName by remember { mutableStateOf("") }
-            TextField(
-                value = uiState.userName,
-                onValueChange = {userName ->  uiState = uiState.copy(userName = userName)},
-                label = { Text(text = "Username") }
+                label = { Text(text = "Email") },
+                onValueChange = {email ->  uiState = uiState.copy(email = email)}
             )
 
             Spacer(modifier = Modifier.height(20.dp))
             //var password by remember { mutableStateOf("") }
             TextField(
                 value = uiState.password,
-                onValueChange = {password -> uiState = uiState.copy(password = password) },
-                label = { Text(text = "Password") }
+                label = { Text(text = "Password") },
+                onValueChange = {password -> uiState = uiState.copy(password = password)}
             )
+
             Spacer(modifier = Modifier.height(1.dp))
 
             ClickableText(
@@ -175,19 +139,18 @@ class SignInActivity : ComponentActivity() {
 
                 )
             Spacer(modifier = Modifier.height(25.dp))
-
             Box(modifier = Modifier.padding(40.dp, 0.dp, 40.dp, 0.dp)) {
                 Button(
                     colors = ButtonDefaults.outlinedButtonColors(backgroundColor = Color.Blue),
                     onClick = {
-                        createAccount()
+                        login()
                     },
                     shape = RoundedCornerShape(50.dp),
                     modifier = Modifier
                         .width(300.dp)
                         .height(50.dp)
                 ) {
-                    Text(text = "Sign Up Free!", style = TextStyle(fontSize = 25.sp), color = Color.White)
+                    Text(text = "Login", style = TextStyle(fontSize = 25.sp), color = Color.White)
                 }
             }
             Spacer(modifier = Modifier.height(20.dp))
@@ -198,7 +161,7 @@ class SignInActivity : ComponentActivity() {
                         fontSize = 14.sp
                     )
                 ) {
-                    append("Have an account? ")
+                    append("No account? ")
                 }
                 withStyle(
                     style = SpanStyle(
@@ -206,15 +169,13 @@ class SignInActivity : ComponentActivity() {
                         fontSize = 14.sp
                     )
                 ) {
-                    append("Sign in here!")
+                    append("Sign up here!")
                 }
             }
             ClickableText(
                 text = text,
-                onClick = { onSignInClick() },
+                onClick = { onSignUpClick() },
             )
-
         }
-
     }
 }
