+++ app/src/main/java/com/example/chessgo/frontend/registration/sign_up/User.kt
@@ -1,4 +1,4 @@
-package com.example.chessgo.frontend.registration.sign_in
+package com.example.chessgo.frontend.registration.sign_up
 
 data class User(val userName: String? = null, val email: String? = null){
